minor code-base refactoring

+++ dataRetriever/src/main/java/com/example/demo/entities/WeatherEntity.java
@@ -10,9 +10,7 @@ public class WeatherEntity {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long id;
-
-    //private Integer weatherDataId;
+    private Integer id;
 
     @Column(name = "weather_id")
     private Integer weatherId;
