+++ weatherDataService/src/main/java/com/dataservice/weatherDataService/entities/MainDataEntity.java
@@ -2,7 +2,6 @@ package com.dataservice.weatherDataService.entities;
 
 import jakarta.persistence.*;
 import lombok.Data;
-
 @Data
 @Table(name = "main_data_table")
 @Entity
@@ -10,7 +9,7 @@ public class MainDataEntity {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long id;
+    private Integer id;
 
     @Column(name = "weather_id")
     private Integer weatherDataId;
